fix(app/businessOrder)修改了order的部分字段,部分重写了businessOrder

+++ pkg/models/order_detail.go
@@ -6,8 +6,8 @@ type OrderDetail struct {
 	gorm.Model
 	Name      string  `gorm:"comment '商品名称'"`
 	Image     string  `gorm:"comment '商品图片'"`
-	OrderId   int     `gorm:"comment '订单id'"`
-	ProductId int     `gorm:"comment '商品id'"`
-	Number    int     `gorm:"comment '商品数量'"`
-	Amount    float64 `gorm:"type:decimal(10,2) comment '支付金额,精确到后两位'"`
+	OrderId   uint32  `gorm:"comment '订单id'"`
+	ProductId uint32  `gorm:"comment '商品id'"`
+	Number    uint32  `gorm:"comment '商品数量'"`
+	Amount    float32 `gorm:"type:decimal(10,2) comment '支付金额,精确到后两位'"`
 }
